+++ posthog-android/src/main/java/com/posthog/android/replay/internal/ViewTreeSnapshotStatus.kt
@@ -1,10 +1,12 @@
 package com.posthog.android.replay.internal
 
+import android.view.ViewTreeObserver
 import com.posthog.internal.replay.RRWireframe
 
 // if you add any new property, remember to clear the state from resetViewSnapshotStates
 internal class ViewTreeSnapshotStatus(
     val listener: NextDrawListener,
+    val layoutListener: ViewTreeObserver.OnGlobalLayoutListener? = null,
     var sentFullSnapshot: Boolean = false,
     var sentMetaEvent: Boolean = false,
     var keyboardVisible: Boolean = false,